services/adk-agent/requirements.txt
google-adk
google-generativeai
toolbox-core
uvicorn
fastapi
